+++ go/parser.go
@@ -2277,7 +2277,7 @@ func (ctxt *parseContext) lookahead_0(initialLine *Line) bool {
         match = true;
         break
       }
-      if false  || ctxt.isMatch_Empty(line) || ctxt.isMatch_Comment(line) || ctxt.isMatch_TagLine(line) {
+      if !(false  || ctxt.isMatch_Empty(line) || ctxt.isMatch_Comment(line) || ctxt.isMatch_TagLine(line)) {
         break
       }
       if atEof {