+++ python/gherkin3/ast_builder.py
@@ -35,8 +35,7 @@ class AstBuilder(object):
         return self.stack[-1]
 
     def get_location(self, token, column=None):
-        # TODO: translated from JS... is it right?
-        return (token.location if (not column or column == 0) else
+        return (token.location if not column else
                 {'line': token.location['line'], 'column': column})
 
     def get_tags(self, node):
@@ -98,7 +97,7 @@ class AstBuilder(object):
             })
         elif node.rule_type == 'DocString':
             separator_token = node.get_tokens('DocStringSeparator')[0]
-            content_type = separator_token.matched_text
+            content_type = separator_token.matched_text if len(separator_token.matched_text) > 0 else None 
             line_tokens = node.get_tokens('Other')
             content = '\n'.join([t.matched_text for t in line_tokens])